Ready for DB implementation.

+++ app/src/main/java/com/example/nav_components_2_tabs_exercise/screens/main/tabs/dashboard/BoxFragment.kt
@@ -51,7 +51,7 @@ class BoxFragment: Fragment(R.layout.fragment_box) {
         }
 
 
-    private fun getBoxId(): Int = args.boxID
+    private fun getBoxId(): Long = args.boxID
 
     private fun getColorValue(): Int = args.boxColorValue
